LayoutTests/http/tests/security/seamless/seamless-sandbox-srcdoc-expected.txt

PASS iframe.seamless is true
PASS window.getComputedStyle(iframe).width is "200px"
PASS window.getComputedStyle(iframe).height is "100px"
